s3_templates/mlops-github-actions/lambda_functions/requirements.txt
boto3
pygithub